+++ branch-master/sqlmath_wasm.js
@@ -1,7 +1,7 @@
 /*jslint-disable*/
 // Copyright (c) 2021 Kai Zhu
 // SPDX-License-Identifier: MIT
-// 2026-05-02T22:27:16+0000
+// 2026-06-22T16:58:19+0000
 (function () {
 "use strict";